break out confusables to security.ex
